refactor(users): reorganize profile endpoints into separate files
- Split profile endpoints into separate files for better modularity
- Move user profile endpoints from auth.py to profile/ directory
- Create constants module for shared application constants
- Update authentication error messages with Soviet-themed language
- Change locked account status code to 410 GONE
- Add test to verify profile router structure
