+++ backend/src/backend/routes/users/__init__.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter
 
-from backend.routes.users import auth
-from backend.routes.users import profile
+from backend.routes.users.auth import router as auth_router
+from backend.routes.users.profile import router as profile_router
 
 router = APIRouter()
 
-router.include_router(auth.router, prefix="/auth")
-router.include_router(profile.router, prefix="/profile")
+router.include_router(auth_router)
+router.include_router(profile_router)
