+++ backend/src/backend/utils/constants.py
@@ -1,0 +1,6 @@
+"""Shared constants used throughout the application."""
+
+from typing import Final
+
+# Marker for unknown IP addresses in request logs
+UNKNOWN_IP_ADDRESS_MARKER: Final[str] = "<UNKNOWN_IP_ADDRESS>"
